Fixing type warnings in LSCompiler

+++ src/module/ls/LSCompiler.h
@@ -14,10 +14,10 @@ class LSCompiler {
 protected:
   LSParser _parser;
   struct Function {
-    L::Table<L::Symbol, uint32_t> local_table;
+    L::Table<L::Symbol, uint8_t> local_table;
     L::Var code;
     L::Array<L::ScriptInstruction> bytecode;
-    uint32_t local_count;
+    uint8_t local_count;
     uint32_t bytecode_offset;
   };
   L::Array<Function*> _functions;
@@ -37,14 +37,14 @@ public:
 protected:
   Function& make_function(const L::Var&);
   void resolve_locals(Function&, const L::Var&);
-  void compile(Function&, const L::Var&, uint32_t offset = 0);
+  void compile(Function&, const L::Var&, uint8_t offset = 0);
   void compile_function(Function&);
   //! Puts object at offset+1 and last index at offset
   //! If get is true, does the final access to put the value at offset
-  void compile_access_chain(Function&, const L::Array<L::Var>& array, uint32_t offset, bool get);
-  void compile_function_call(Function& func, const L::Array<L::Var>& array, uint32_t offset);
-  void compile_assignment(Function& func, const L::Var& dst, uint8_t src, uint32_t offset);
-  void compile_operator(Function& func, const L::Array<L::Var>& array, uint32_t offset, L::ScriptOpCode opcode);
-  void compile_op_assign(Function& func, const L::Array<L::Var>& array, uint32_t offset, L::ScriptOpCode opcode);
-  void compile_comparison(Function& func, const L::Array<L::Var>& array, uint32_t offset, L::ScriptOpCode cmp, bool not = false);
+  void compile_access_chain(Function&, const L::Array<L::Var>& array, uint8_t offset, bool get);
+  void compile_function_call(Function& func, const L::Array<L::Var>& array, uint8_t offset);
+  void compile_assignment(Function& func, const L::Var& dst, uint8_t src, uint8_t offset);
+  void compile_operator(Function& func, const L::Array<L::Var>& array, uint8_t offset, L::ScriptOpCode opcode);
+  void compile_op_assign(Function& func, const L::Array<L::Var>& array, uint8_t offset, L::ScriptOpCode opcode);
+  void compile_comparison(Function& func, const L::Array<L::Var>& array, uint8_t offset, L::ScriptOpCode cmp, bool not = false);
 };
